refactor(core): extraer helpers de superadmins, team y onboarding
- Crear superadmins/_helpers.tsx con formatDate, isInactive,
  extractErrorMessage, statusBadge, roleBadge, ROLE_LABELS,
  AUDIT_ACTION_LABELS, CreateFormState/EMPTY_FORM
- Crear team/_helpers.tsx con PROVIDER_CONFIG, ROLE_CONFIG,
  getInitials, isValidAvatarUrl, ProviderBadge, AuthMethodBadge
- Crear onboarding/_helpers.ts con SECTIONS, session helpers
  (saveToSession, loadFromSession, clearSession),
  FORMAT_VALIDATORS, getDefaultSections
- Reducir superadmins/page.tsx de 1355 a 1261 lineas (-94)
- Reducir team/page.tsx de 1041 a 964 lineas (-77)
- Reducir onboarding/page.tsx de 1167 a 1086 lineas (-81)

Refs: #244
