+++ frontend/src/app/(tenant)/dashboard/settings/team/page.tsx
@@ -64,12 +64,8 @@ import {
   Users,
   MoreVertical,
   Unlink,
-  Shield,
   ShieldCheck,
   ShieldOff,
-  UserRound,
-  Mail,
-  KeyRound,
   UserPlus,
   Ban,
   UserCheck,
@@ -84,90 +80,17 @@ import {
   Info,
 } from 'lucide-react';
 import { toast } from 'sonner';
-import Image from 'next/image';
 import { useRouter } from 'next/navigation';
-
-// ─── Estilos hoisted (consistentes con profile) ──────────
-const navyText = { color: '#1C3B57' } as const;
-const navyIconBg = { background: 'rgba(28, 59, 87, 0.06)' } as const;
-
-const PROVIDER_CONFIG = {
-  google: { label: 'Google', textColor: 'text-blue-700', bgLight: 'bg-blue-50' },
-  apple: { label: 'Apple', textColor: 'text-neutral-900', bgLight: 'bg-neutral-100' },
-  facebook: { label: 'Facebook', textColor: 'text-indigo-700', bgLight: 'bg-indigo-50' },
-} as const;
-
-const ROLE_CONFIG = {
-  admin: { label: 'Admin', variant: 'default' as const, icon: ShieldCheck, color: 'text-emerald-600', bg: 'bg-emerald-500/10' },
-  staff: { label: 'Staff', variant: 'secondary' as const, icon: Shield, color: 'text-amber-600', bg: 'bg-amber-500/10' },
-  customer: { label: 'Cliente', variant: 'outline' as const, icon: UserRound, color: 'text-blue-600', bg: 'bg-blue-500/10' },
-} as const;
-
-function getInitials(member: TeamMember): string {
-  if (member.first_name && member.last_name) {
-    return `${member.first_name[0]}${member.last_name[0]}`.toUpperCase();
-  }
-  return member.email[0].toUpperCase();
-}
-
-const DEFAULT_PROVIDER_CONFIG = { label: 'Otro', textColor: 'text-gray-700', bgLight: 'bg-gray-100' };
-
-function isValidAvatarUrl(url: string): boolean {
-  try {
-    const parsed = new URL(url);
-    return parsed.protocol === 'https:' || parsed.protocol === 'http:';
-  } catch {
-    return false;
-  }
-}
-
-function ProviderBadge({ social }: { social: SocialAccountDetail }) {
-  const config = PROVIDER_CONFIG[social.provider] ?? DEFAULT_PROVIDER_CONFIG;
-  const showAvatar = social.avatar_url && isValidAvatarUrl(social.avatar_url);
-  return (
-    <span
-      className={`inline-flex items-center gap-1.5 px-2 py-0.5 rounded-full text-xs font-medium ${config.bgLight} ${config.textColor}`}
-    >
-      {showAvatar && (
-        <Image
-          src={social.avatar_url}
-          alt=""
-          width={14}
-          height={14}
-          className="w-3.5 h-3.5 rounded-full"
-          unoptimized
-        />
-      )}
-      {config.label}
-    </span>
-  );
-}
-
-function AuthMethodBadge({ method }: { method: TeamMember['auth_method'] }) {
-  if (method === 'email_only') {
-    return (
-      <span className="inline-flex items-center gap-1 text-xs text-muted-foreground">
-        <Mail className="w-3 h-3" />
-        Email
-      </span>
-    );
-  }
-  if (method === 'social_only') {
-    return (
-      <span className="inline-flex items-center gap-1 text-xs text-muted-foreground">
-        <KeyRound className="w-3 h-3" />
-        Social
-      </span>
-    );
-  }
-  return (
-    <span className="inline-flex items-center gap-1 text-xs text-muted-foreground">
-      <Mail className="w-3 h-3" />
-      +
-      <KeyRound className="w-3 h-3" />
-    </span>
-  );
-}
+import {
+  navyText,
+  navyIconBg,
+  PROVIDER_CONFIG,
+  ROLE_CONFIG,
+  DEFAULT_PROVIDER_CONFIG,
+  getInitials,
+  ProviderBadge,
+  AuthMethodBadge,
+} from './_helpers';
 
 // ─── Página de equipo ─────────────────────────────────────
 export default function SettingsTeamPage() {
